requirements.txt
# Trading Bot Requirements

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.14.0

# Claude AI
anthropic>=0.18.0

# HTTP requests (for Hyperliquid API + Telegram)
requests>=2.31.0

# Hyperliquid SDK
hyperliquid-python-sdk>=0.19.0
eth-account>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Telegram notifications
# python-telegram-bot>=20.0

# Optional: Advanced backtesting
# backtrader>=1.9.76
